backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pandas>=2.2.2
duckdb>=1.0.0
openai>=1.37.0
pydantic>=2.7.1
orjson>=3.10.5
python-dotenv>=1.0.1
openpyxl>=3.1.2

